requirements.txt
google-generativeai>=0.8.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
Pillow>=10.0
pillow-heif>=0.16
tqdm>=4.66
click>=8.1
imagehash>=4.3
thefuzz[speedup]>=0.20
structlog>=24.0
